+++ inst/tinytest/test_is_natural.R
@@ -1,3 +1,15 @@
+#### Preparations ####
+toy_fun_erroneous <- function(x) {
+  stopifnot(is_natural(x))
+  seq_len(x)
+}
+toy_fun_correct <- function(x) {
+  stopifnot(is_natural(x))
+  x <- round(x)
+  seq_len(x)
+}
+
+
 #### Test the examples ####
 expect_true(is_natural(x = 5 + 1e-10))
 expect_false(is_natural(x = 1e-10))
@@ -21,6 +33,10 @@ expect_false(all_natural(x = c(3, 5, Inf), strict = FALSE))
 expect_false(all_natural(x = "a"))
 expect_true(all_natural(x = 1:2))
 
+expect_equal(toy_fun_erroneous(x = 5 - 1e-8), 1:4)
+expect_equal(toy_fun_correct(x = 5 - 1e-8), 1:5)
+
+
 #### Tests ####
 expect_false(is_natural(x = -3L, strict = FALSE))
 expect_false(is_natural(x = 0, strict = TRUE))
@@ -29,10 +45,16 @@ expect_false(is_natural(x = 0L, strict = TRUE))
 expect_true(is_natural(x = 0L, strict = FALSE))
 expect_false(is_natural(x = Inf, strict = TRUE))
 expect_false(is_natural(x = Inf, strict = FALSE))
+expect_false(is_natural(x = integer(0), allow_zero = FALSE, strict = FALSE))
+expect_true(is_natural(x = integer(0), allow_zero = TRUE, strict = FALSE))
+expect_false(is_natural(x = numeric(0), allow_zero = FALSE, strict = FALSE))
+expect_true(is_natural(x = numeric(0), allow_zero = TRUE, strict = FALSE))
+expect_false(is_natural(x = NULL, allow_zero = TRUE, strict = FALSE))
 
 # Testing a value that cannot be represented as integer by R (R uses 32-bit
-# integers, see 'Details' in help(`integer`))
-expect_true(is_natural(x = .Machine$integer.max + 10)) #
+# integers, see 'Details' in help(`integer`)). Negative values are not natural
+# anyway, so no need to test large negative numbers.
+expect_false(is_natural(x = .Machine$integer.max + 1))
 
 expect_false(is_natural(x = c(3, 5 + 1e-10, 5.0, 1e4, 1.2e4)))
 
@@ -59,10 +81,16 @@ expect_false(all_natural(x = 0L, strict = TRUE))
 expect_true(all_natural(x = 0L, strict = FALSE))
 expect_false(all_natural(x = Inf, strict = TRUE))
 expect_false(all_natural(x = Inf, strict = FALSE))
+expect_false(all_natural(x = integer(0), allow_zero = FALSE, strict = FALSE))
+expect_true(all_natural(x = integer(0), allow_zero = TRUE, strict = FALSE))
+expect_false(all_natural(x = numeric(0), allow_zero = FALSE, strict = FALSE))
+expect_true(all_natural(x = numeric(0), allow_zero = TRUE, strict = FALSE))
+expect_false(all_natural(x = NULL, allow_zero = TRUE, strict = FALSE))
 
 # Testing a value that cannot be represented as integer by R (R uses 32-bit
-# integers, see 'Details' in help(`integer`))
-expect_true(all_natural(x = .Machine$integer.max + 10)) #
+# integers, see 'Details' in help(`integer`)). Negative values are not natural
+# anyway, so no need to test large negative numbers.
+expect_false(all_natural(x = .Machine$integer.max + 1))
 
 expect_true(all_natural(x = c(3, 5 + 1e-10, 5.0, 1e4, 1.2e4)))
 
@@ -102,4 +130,4 @@ expect_error(all_natural(x = c(3, 5 + 1e-10), tol = 0),
 
 
 #### Remove objects used in tests ####
-rm(allow_NA, x)
+rm(allow_NA, toy_fun_correct, toy_fun_erroneous, x)